Add a .mailmap file
So I was checking out minetest and wondered how many developers there are
and how they development community is setup and such, so I decided to have
a look at the git shortlog. There were lot of double entries for some
persons, so it was harder to estimate the number of people involved.

By adding a .mailmap file, the overview of the authors is much improved
For details on the .mailmap file, see
https://www.kernel.org/pub/software/scm/git/docs/git-shortlog.html

Signed-off-by: Stefan Beller <stefanbeller@gmail.com>